Rename withTeamStore file
